[fix][misc] Move Dock spring-loading command to misc
**Why this change was necessary**
Since this command's behaviour is very subtle, it's been moved to the
`misc` folder. The app-specific subtlety of this command was not previously described.

**What this change does**
See title. Further detail is added about the command relying on an
app being open as well as apps this command has been observed to affect.

**Any side-effects?**
None

**Additional context/notes/links**

Resolves yannbertrand/macos-defaults#51
Related to https://github.com/yannbertrand/macos-defaults/pull/52#issuecomment-703279333