[feat][clock] Add videos
**Why this change was necessary**
The video recordings of the different command parameters have to be
added manually to the repo. The filenames had to be modified to
produce colon-free filenames.

**What this change does**
Adds video recordings of different command parameters. Changes
filenames of recordings and the scripts that generate them.

**Any side-effects?**
None.

**Additional context/notes/links**

Resolves yannbertrand/macos-defaults#41
Related to https://github.com/yannbertrand/macos-defaults/pull/56#issuecomment-703301979