[feat][TimeMachine] Add command for new disk backup prompt
**Why this change was necessary**
By default, when a new disk is connected to a Mac for the first time
after it's been formatted, Time Machine asks the user if they want to
add it as a backup volume. This can be tedious for users that
frequently use and format portable hard drives for data transfer.

**What this change does**
Adds a defaults command allowing the user to disable these prompts
when a new disk is connected.

**Any side-effects?**
None.

**Additional context/notes/links**

Resolves yannbertrand/macos-defaults#54