Improve markdown support
This improves markdown support within descriptions. Currently, adding
multiline text blocks in the defaults.yml file breaks, because the
generated head meta tags content includes that whitespace as is and
breaks the generated yaml contents. Now, meta tag content is stringified
as JSON, which is yaml compliant. The final html meta tag contains the
markdown as is. Also currently, markdown descriptions inserted into the
content can interact with subsequent content (if you end with a list
it'll cause funky interaction with the following content). That's fixed
by inserting a "break" between (done on categores and pages).